Organize expansion cards by set
