+++ dominion/cards/adventures/__init__.py
@@ -1,0 +1,3 @@
+from .miser import Miser
+
+__all__ = ['Miser']
