+++ dominion/cards/allies/__init__.py
@@ -1,0 +1,4 @@
+from .collection import Collection
+from .modify import Modify
+
+__all__ = ['Collection', 'Modify']
