+++ dominion/cards/empires/__init__.py
@@ -1,0 +1,4 @@
+from .patrician import Patrician
+from .emporium import Emporium
+
+__all__ = ['Patrician', 'Emporium']
