+++ dominion/cards/promo/__init__.py
@@ -1,0 +1,3 @@
+from .snowy_village import SnowyVillage
+
+__all__ = ['SnowyVillage']
